--- src/main/java/com/pettrip/app/dto/ResponseDTO.java
@@ -1,15 +1,0 @@
-package com.pettrip.app.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.context.annotation.Bean;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class ResponseDTO {
-    private boolean isSuccess;
-    private int code;
-    private String message;
-}
