+++ src/main/java/com/pettrip/domain/User.java
@@ -1,12 +1,11 @@
 package com.pettrip.domain;
 
 import lombok.*;
-import com.pettrip.app.dto.JoinDTO;
-import com.pettrip.app.dto.LoginDTO;
+import com.pettrip.app.dto.user.JoinDTO;
+import com.pettrip.app.dto.user.LoginDTO;
 import com.pettrip.domain.common.BaseEntity;
 
 import javax.persistence.*;
-import java.util.List;
 
 @Entity
 @Getter
